// clients/ts/src/generated/accounts/config.ts
export const CONFIG_DISCRIMINATOR: ReadonlyUint8Array = new Uint8Array([
  155, 12, 170, 224, 30, 250, 204, 130,
]);

export function getConfigDiscriminatorBytes(): ReadonlyUint8Array {
  return fixEncoderSize(getBytesEncoder(), 8).encode(CONFIG_DISCRIMINATOR);
}

export type Config = {
  discriminator: ReadonlyUint8Array;
  authority: Address;
  allowedMint: Address;
  tokenProgram: Address;
  safetyWindowSeconds: bigint;
  claimWindowSeconds: bigint;
  privateValidator: Address;
  version: number;
  bump: number;
};